1. Provision certificate chain with out using extra RAM. 2. Made changes to support only EC signing key instead of RSA key. 3. Clear HMAC nonce and generate it in setBootParams.

+++ Applet/Applet/AndroidSEProvider/com/android/javacard/keymaster/KMRsaOAEPEncoding.java
@@ -170,7 +170,7 @@ public class KMRsaOAEPEncoding extends Cipher {
     try {
       short hLen = getDigestLength();
 
-      if (encodedMsgLen < (2 * hLen + 1)) {
+      if (encodedMsgLen < (short)(2 * hLen + 1)) {
         CryptoException.throwIt(CryptoException.ILLEGAL_VALUE);
       }
       // encodedMsg will be in the format of maskedSeed||maskedDB.
